requirements.txt
transformers==4.41.2
datasets==2.19.2
peft==0.11.1
accelerate==0.31.0
modelscope>=1.35,<2
tqdm>=4.66
numpy>=1.24
pyyaml>=6.0
matplotlib>=3.8
